Move effects to individual files
