+++ src/resources.rs
@@ -1,4 +1,4 @@
 pub mod color;
-pub mod effect;
+pub mod effects;
 pub mod ledstrip;
 pub mod settings;
